# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: string;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface Conversation {
  id: string;
  title: string;
  preview: string;
  updatedAt: string;
  messageCount: number;
}

export type ConversationDetail = Conversation & {
  messages: Message[];
};

export interface ChatInput {
  /** @nullable */
  conversationId?: string | null;
  /** @minLength 1 */
  prompt: string;
}

export interface ChatResponse {
  conversation: ConversationDetail;
  assistantMessage: Message;
}