# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatInput {
  /** @nullable */
  conversationId?: string | null;
  /** @minLength 1 */
  prompt: string;
}