# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationDetail } from './conversationDetail';
import type { Message } from './message';

export interface ChatResponse {
  conversation: ConversationDetail;
  assistantMessage: Message;
}