# lib/api-zod/src/generated/types/conversationDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Conversation } from './conversation';
import type { Message } from './message';

export type ConversationDetail = Conversation & {
  messages: Message[];
};